generate secret proxy key dynamically
Instead of having to “spoil” the config with a secret key (which hinders
uploading to a public repo), we can generate it on application start and put it
into the user cache directory (which should only be readable by the searx user).

Adds `utils.get_secret_app_key` which generates a key if none exists and returns
the existing one otherwise (declarative style).
Also test every codepath of that function.