[enh] Add timeout limit per request
The new url parameter "timeout_limit" set timeout limit defined in second.
Example "timeout_limit=1.5" means the timeout limit is 1.5 seconds

In addition, the query can start with <[number] to set the timeout limit.

For number between 0 and 9, the number unit is the second :
Example: "<3 searx" means the timeout limit is 3 seconds

For number above 9, the number unit is the centisecond:
Example: "<150 searx" means the timeout is 1.5 seconds

Related to #1077
Updated version of PR #1413 from @isj-privacore