Proposal: logicodev follow color scheme from OS
This is a suggestion/proposal to make the default theme follow the color
scheme setting of the OS or browser. With the commit applied, you have
three logicodev variants:
logicodev: Automatically follows dark/light mode as set in OS/browser,
and falls back to light if no setting is found.
logicodev-dark: always dark
logicodev-light: always light