Split autocompleter into (browser)suggestions and API
Since current browers do not seem to either POST requests at all, or
supplying a format-parameter as the current assumption is, the easiest
reasonably correct way seems to be to split the autocompleter endpoint
into:
- /autocompleter as used currently, continuing to work for the API-usage
  on the web-frondend
- /autocompleter_suggestions which is included in opensearch.xml defined
  for browser-suggestios-requests

While it is sad no get rid of POST for browser-autocompleter, it seems
neither chromium nor firefox currently support it — and opensearch
cannot reasonably specify it, as far as I understand. Thus removing it
seems fair for the time being.

Fixes: #2105